paso toda la logica de los botones para editar y eliminar al useActivities asi el componente queda mas limpio
