+++ src/Components/views/Private/BackOffice/Acivities/List/ActivitiesList.jsx
@@ -1,62 +1,16 @@
-import { useState } from "react";
-
 import { addIcon } from "assets/images";
 import { BackTo } from "Components/GlobalComponents/BackTo/BackTo";
 import { CustomAlertMessage } from "Components/GlobalComponents/CustomAlertMessage/CustomAlertMessage";
 import { CustomTable } from "Components/GlobalComponents/CustomTable/CustomTable";
 import { CustomTitle } from "Components/GlobalComponents/CustomTitle/CustomTitle";
-import { feedbackUser } from "utilities/alerts/feedbackUser.util";
-import { handleUserConfirm as AlertWarning } from "utilities/alerts/userConfirm.util";
 import { privateRoutes } from "models/routes";
-import privateService from "Services/privateApiService";
 import SearchActivities from "./components/SearchActivities";
-import { URLs } from "Services/ServicesURLS";
 import { useActivities } from "./hooks/useActivities";
-import { useNavigate } from "react-router-dom";
-import { errorMessages } from "../utilities/errorMessages";
-import { successMessages } from "../utilities/successMessages";
 
 const ActivitiesList = () => {
   const tHead = ["#", "Nombre", "Imagen", "Fecha de Creación", "Acciones"];
   const myTableData = { name: "name", image: "image", created_at: "created_at" };
-  const { activities, loadingActivities, activitiesData, fetchActivities } = useActivities();
-  const [loading, setLoading] = useState(false);
-  const navigate = useNavigate();
-
-  const editHandler = (id) => {
-    navigate(`/${privateRoutes.BACKOFFICE}${privateRoutes.ACTIVITIESEDIT}/${id}`);
-  };
-
-  const deleteHandler = async (id) => {
-    const find = activitiesData.find((el) => id === el.id);
-
-    if (find) {
-      const response = await AlertWarning("¿Estas segura/o que deseas eliminar la actividad?");
-
-      if (response) {
-        setLoading(true);
-        const activityDeleted = await privateService.deleted(URLs.activity, id);
-
-        if (activityDeleted) {
-          await feedbackUser("center", "success", `${successMessages.deleteActivity}`);
-          fetchActivities();
-        } else {
-          await feedbackUser("center", "error", `${errorMessages.deleteActivity}`);
-        }
-      }
-      setLoading(false);
-    }
-  };
-
-  const searchActivitiesHandler = async (searchText) => {
-    const fetchParams = {};
-
-    if (searchText.length >= 3) {
-      fetchParams["search"] = searchText;
-    }
-
-    await fetchActivities(fetchParams);
-  };
+  const { activities, loadingActivities, activitiesData, loading, editHandler, deleteHandler, searchActivitiesHandler } = useActivities();
 
   let activitiesContent;
   if (loadingActivities || loading) {
