+++ src/Components/views/Private/BackOffice/Acivities/List/hooks/useActivities.js
@@ -1,16 +1,58 @@
 import { useEffect, useState } from "react";
+import { useNavigate } from "react-router-dom";
 
-import { feedbackUser } from "utilities/alerts/feedbackUser.util";
-import { URLs } from "Services/ServicesURLS";
-import privateService from "Services/privateApiService";
 import { encodeQueryParams } from "utilities/queryParams";
 import { errorMessages } from "../../utilities/errorMessages";
+import { feedbackUser } from "utilities/alerts/feedbackUser.util";
+import { handleUserConfirm as AlertWarning } from "utilities/alerts/userConfirm.util";
+import { privateRoutes } from "models/routes";
+import privateService from "Services/privateApiService";
+import { successMessages } from "../../utilities/successMessages";
+import { URLs } from "Services/ServicesURLS";
 import { whatIs } from "utilities/parseDate";
 
 export const useActivities = () => {
   const [activitiesData, setActivitiesData] = useState([]);
   const [activities, setActivities] = useState([]);
   const [loadingActivities, setLoadingActivities] = useState(true);
+  const [loading, setLoading] = useState(false);
+
+  const navigate = useNavigate();
+
+  const editHandler = (id) => {
+    navigate(`/${privateRoutes.BACKOFFICE}${privateRoutes.ACTIVITIESEDIT}/${id}`);
+  };
+
+  const deleteHandler = async (id) => {
+    const find = activitiesData.find((el) => id === el.id);
+
+    if (find) {
+      const response = await AlertWarning("¿Estas segura/o que deseas eliminar la actividad?");
+
+      if (response) {
+        setLoading(true);
+        const activityDeleted = await privateService.deleted(URLs.activity, id);
+
+        if (activityDeleted) {
+          await feedbackUser("center", "success", `${successMessages.deleteActivity}`);
+          fetchActivities();
+        } else {
+          await feedbackUser("center", "error", `${errorMessages.deleteActivity}`);
+        }
+      }
+      setLoading(false);
+    }
+  };
+
+  const searchActivitiesHandler = async (searchText) => {
+    const fetchParams = {};
+
+    if (searchText.length >= 3) {
+      fetchParams["search"] = searchText;
+    }
+
+    await fetchActivities(fetchParams);
+  };
 
   const fetchActivities = async (params = {}) => {
     try {
@@ -36,5 +78,5 @@ export const useActivities = () => {
     fetchActivities();
   }, []);
 
-  return { activities, loadingActivities, activitiesData, fetchActivities };
+  return { activities, loadingActivities, activitiesData, loading, editHandler, deleteHandler, searchActivitiesHandler, fetchActivities };
 };
